feat: 댓글 수정 및 삭제 권한 추가

+++ freelec-springboot2-webservice/src/main/resources/static/js/app/index.js
@@ -117,13 +117,22 @@ var main = { //main이란 객체 생성해서 function 유효 범위 선언
             });
         }
     },
-    <!--comment update 기능 추가-->
     commentUpdate : function (form) {
         const data = {
             id: form.querySelector('#id').value,
             postsId: form.querySelector('#postsId').value,
             comment: form.querySelector('#comment-content').value,
+            writerUserId: form.querySelector('#writerUserId').value,
+            sessionUserId: form.querySelector('#sessionUserId').value
         }
+        console.log("commentWriterID : " + data.writerUserId);
+        console.log("sessionUserID : " + data.sessionUserId);
+
+        if (data.writerUserId !== data.sessionUserId) {
+            alert("본인이 작성한 댓글만 수정 가능합니다.");
+            return false;
+        }
+
         if (!data.comment || data.comment.trim() === "") {
             alert("공백 또는 입력하지 않은 부분이 있습니다.");
             return false;
@@ -132,7 +141,7 @@ var main = { //main이란 객체 생성해서 function 유효 범위 선언
         if (con_check === true) {
             $.ajax({
                 type: 'PUT',
-                url: '/api/v1/posts/' + data.postsId + '/comments/' + data.id,
+                url: '/api/posts/' + data.postsId + '/comments/' + data.id,
                 dataType: 'JSON',
                 contentType: 'application/json; charset=utf-8',
                 data: JSON.stringify(data)
@@ -143,21 +152,30 @@ var main = { //main이란 객체 생성해서 function 유효 범위 선언
             });
         }
     },
-    <!--comment delete 기능 추가-->
-    commentDelete : function (postsId, commentId) {
-        const con_check = confirm("삭제하시겠습니까?");
-        if (con_check === true) {
-            $.ajax({
-                type: 'DELETE',
-                url: '/api/v1/posts/' + postsId + '/comments/' + commentId,
-                dataType: 'JSON',
-            }).done(function () {
-                alert('댓글이 삭제되었습니다.');
-                window.location.reload();
-            }).fail(function (error) {
-                alert(JSON.stringify(error));
-            });
+
+    /** 댓글 삭제 */
+    commentDelete : function (postsId, commentId, commentWriterId, sessionUserId) {
+
+        // 본인이 작성한 글인지 확인
+        if (commentWriterId !== sessionUserId) {
+            alert("본인이 작성한 댓글만 삭제 가능합니다.");
+        } else {
+            const con_check = confirm("삭제하시겠습니까?");
+
+            if (con_check === true) {
+                $.ajax({
+                    type: 'DELETE',
+                    url: '/api/posts/' + postsId + '/comments/' + commentId,
+                    dataType: 'JSON',
+                }).done(function () {
+                    alert('댓글이 삭제되었습니다.');
+                    window.location.reload();
+                }).fail(function (error) {
+                    alert(JSON.stringify(error));
+                });
+            }
         }
     }
+
 };
 main.init();
